requirements.txt
paho-mqtt~=1.5.1
inverterd~=1.0.3
clickhouse-driver~=0.2.0
toml~=0.10.2
Flask~=2.0.2
mysql-connector-python~=8.0.27
Werkzeug~=2.0.2
uwsgi~=2.0.20
python-telegram-bot~=13.1
inverterd~=1.0.2
requests~=2.26.0
aiohttp~=3.8.1
pytz~=2021.3

# following can be installed from debian repositories
# matplotlib~=3.5.0

